all solvers now return the same Result object
- Renamed `ComplexResult` to `Result`
- Refactored solvers
  - Moved none user facing solver structure into `_base` package and
    separated into their own files.
  - Enumerations have been moved into their own files.
  - All solvers will not return an instance of `Result`.
- Refactored `solve` action to reflect changes to solver API

Resolves #53
