+++ python/src/solvers/_base/__init__.py
@@ -1,0 +1,5 @@
+from .enumerations import IterativeInitialGuess, ResultAttributes
+from .result import Result
+
+from .basic_solver import _BasicSolver
+from .iterative_solver import _IterativeSolver
